# src/Loaders/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SJ CUSTOMS API",
    "version": "1.0.0",
    "description": "API para la gestión de tuneo y personalización de coches"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de desarrollo"
    }
  ],
  "paths": {
...
    "/api/car": {
      "get": {
        "summary": "Obtiene todos los coches",
        "tags": [
          "Car"
        ],
        "responses": {
          "200": {
            "description": "Lista de coches"
          }
        }
      },
      "post": {
        "summary": "Agrega un nuevo coche",
        "tags": [
          "Car"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "brand": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "engineOptions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "imageUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Coche agregado"
          }
        }
      }
    },
    "/api/car/{id}": {
      "get": {
        "summary": "Obtiene un coche por ID",
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coche encontrado"
          },
          "404": {
            "description": "Coche no encontrado"
          }
        }
      },
      "put": {
        "summary": "Actualiza un coche por ID",
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "brand": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number"
                  },
                  "engineOptions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "imageUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Coche actualizado"
          }
        }
      },
      "delete": {
        "summary": "Elimina un coche por ID",
        "tags": [
          "Car"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coche eliminado"
          }
        }
      }
    },
...
  },
...
}